feat: remove `parseExpectCall` and friends
