+++ src/rules/utils/index.ts
@@ -3,4 +3,3 @@ export * from './detectJestVersion';
 export * from './followTypeAssertionChain';
 export * from './misc';
 export * from './parseJestFnCall';
-export * from './parseExpectCall';
